Corrected Timer_Config structure to utilize a float for the frequency
This commit adds corrections to timer_config so that a wider
range of frequencies can be attained. By allowing floats, there is more room
for a larger scale of frequencies.

+++ include/Timer.h
@@ -36,7 +36,7 @@ extern "C" {
 
     typedef struct TIMER_CONFIG {
         Timer_Type which_timer; ///< The timer to configure
-        uint frequency;
+        float frequency;
         uint pbclk;
         void (*callback); ///< The function to call when the timer is triggered
         boolean enabled; ///< Tells if the timer is enabled
